backend/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.0.1
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
ecdsa==0.19.0
fastapi==0.114.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.8
Mako==1.3.5
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.1
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic_core==2.23.2
pydantic==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.20.0
watchfiles==0.24.0
websockets==13.0.1
